Resume Screening Assistance Project/requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
gradio
openai
tiktoken
python-dotenv
unstructured
pinecone-client
pypdf
sentence-transformers